Finished fun addEnchantToList and added fun editEnchant
Added functionality to addEnchantToList to clear the input fields when the data held inside them has already been used to create an object.

Added functionality to enchantCard which allows specific items in the list to be selected. Which item is selected is now stored in the state holder as var selectedEnchant.

Added editEnchant, which populates the input fields with information from the selected item and removes that item from the list.