Rebuilt the existing UI in preparation for adding new features.
The UI has all the same functionality as the previous UI, but is organized slightly differently (both visually and in the code). Edit and Delete buttons are now on each individual list item. A button to clear all input fields now exists. Non-functional buttons that will be used by upcoming features are now in place.