Added functionality to prevent duplicate enchantments from being added.
When a user tries to add an enchantment that is already present in the list (regardless of capitalization), they are greeted with a dialog asking if they'd like to replace the existing enchantment with what they wrote most recently, or if they'd like to cancel the operation.

If the user chooses to replace the existing entry, the existing entry changes and stays where it was before. If the user chooses to cancel, nothing happens and the input fields stay filled.

Also made it so names that are too long for their space in the list are clipped (to preserve alignment and spacing)