Added wizard tab functionality and began work on the "tools" tab.
When a user clicks on a tab, the wizard moves to that tab. In addition, when a user clicks "next" or "previous", it switches tabs accordingly. "cancel" and "finish" buttons were added to the first and last tabs respectively, but the buttons are currently nonfunctional.

The tools tab was created and basic functionality pertaining to tools was added. Users can now add tools to a tools list. Users cannot assign enchantments to tools yet.

Took a month break and came back to the project not compiling, so I'm commiting my work from a month ago so I can safely mess with files to try and fix dependencies.